+++ pom.xml
@@ -25,9 +25,7 @@
         <seed.version>2.1.0-M1-SNAPSHOT</seed.version>
         <hibernate-validator.version>5.1.3.Final</hibernate-validator.version>
 
-        <compatibility.groupId>org.seedstack.seed</compatibility.groupId>
-        <compatibility.artifactId>seed-validation-support</compatibility.artifactId>
-        <compatibility.version>2.0.0</compatibility.version>
+        <compatibility.skip>true</compatibility.skip>
     </properties>
 
     <build>
