+++ src/it/java/org/seedstack/validation/internal/pojo/DummyServiceFieldValidation.java
@@ -1,0 +1,24 @@
+/**
+ * Copyright (c) 2013-2015, The SeedStack authors <http://seedstack.org>
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+package org.seedstack.validation.internal.pojo;
+
+import org.seedstack.seed.it.ITBind;
+
+import javax.validation.constraints.NotNull;
+
+
+@ITBind
+public class DummyServiceFieldValidation {
+
+    @NotNull
+    private Object param;
+
+    public void doSomethingAwesome(Object param) {
+        this.param = param;
+    }
+}
