--- src/main/java/org/seedstack/seed/validation/api/package-info.java
@@ -1,13 +1,0 @@
-/**
- * Copyright (c) 2013-2015, The SeedStack authors <http://seedstack.org>
- *
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
-/**
- * Provides classes for using Java Bean Validation.
- *
- * @author Pierre Thirouin
- */
-package org.seedstack.seed.validation.api;
