+++ src/main/resources/META-INF/errors/org.seedstack.validation.internal.ValidationErrorCode.properties
