+++ src/main/resources/META-INF/services/io.nuun.kernel.api.Plugin
@@ -1,1 +1,1 @@
-org.seedstack.seed.validation.internal.ValidationPlugin
+org.seedstack.validation.internal.ValidationPlugin
